handle update-default when on default branch
When the current working directory is the default branch we can't fetch
into the ref directly. We could enforce this by using `-u`. But the
fetch command leaves the working directory untouched, which is not what
we want.
In summary, when on the default branch we do a rebase, and a direct
fetch otherwise.